+++ src/package/Console/Commands/Base.php
@@ -15,4 +15,21 @@ class Base extends Command
     {
         $this->info(config('app.name').' '.app('pragmarx.version')->format($format));
     }
+
+    /**
+     * Display the current app version.
+     *
+     * @param string $type
+     * @return bool
+     */
+    public function checkIfCanIncrement($type)
+    {
+        if (app('pragmarx.version')->isInAbsorbMode($type)) {
+            $this->error("Version is in git absorb mode, cannot be incremented");
+
+            return false;
+        }
+
+        return true;
+    }
 }